requirements.txt
asgiref==3.8.1
Django==4.2.11
django-auditlog==3.0.0
django-extensions==3.2.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
